# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FitPilot AI gym management and personal fitness API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardOverview {
  memberCount: number;
  activeMembers: number;
  checkinsToday: number;
  attendanceRate: number;
  attendanceChange: number;
  monthlyRevenue: number;
  revenueChange: number;
  programCount: number;
  classCount: number;
}

export interface AttendancePoint {
  label: string;
  checkins: number;
  capacity: number;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  checkin: 'checkin',
  signup: 'signup',
  payment: 'payment',
  class: 'class',
  program: 'program',
} as const;

export interface Activity {
  id: number;
  type: ActivityType;
  title: string;
  detail: string;
  timestamp: string;
  /** @nullable */
  memberName: string | null;
}

export type MemberPlan = typeof MemberPlan[keyof typeof MemberPlan];


export const MemberPlan = {
  Core: 'Core',
  Pro: 'Pro',
  Elite: 'Elite',
} as const;

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  active: 'active',
  paused: 'paused',
  at_risk: 'at_risk',
} as const;

export interface Member {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  plan: MemberPlan;
  status: MemberStatus;
  joinedAt: string;
  /** @nullable */
  lastVisit: string | null;
  visitsThisMonth: number;
  goal: string;
  avatarColor: string;
}

export type MemberInputPlan = typeof MemberInputPlan[keyof typeof MemberInputPlan];


export const MemberInputPlan = {
  Core: 'Core',
  Pro: 'Pro',
  Elite: 'Elite',
} as const;

export interface MemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 3 */
  email: string;
  phone?: string;
  plan: MemberInputPlan;
  /** @minLength 1 */
  goal: string;
}

export type MemberUpdatePlan = typeof MemberUpdatePlan[keyof typeof MemberUpdatePlan];


export const MemberUpdatePlan = {
  Core: 'Core',
  Pro: 'Pro',
  Elite: 'Elite',
} as const;

export type MemberUpdateStatus = typeof MemberUpdateStatus[keyof typeof MemberUpdateStatus];


export const MemberUpdateStatus = {
  active: 'active',
  paused: 'paused',
  at_risk: 'at_risk',
} as const;

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 3 */
  email?: string;
  phone?: string;
  plan?: MemberUpdatePlan;
  status?: MemberUpdateStatus;
  /** @minLength 1 */
  goal?: string;
}

export type ProgramLevel = typeof ProgramLevel[keyof typeof ProgramLevel];


export const ProgramLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Program {
  id: number;
  name: string;
  description: string;
  level: ProgramLevel;
  weeks: number;
  sessionsPerWeek: number;
  activeMembers: number;
  color: string;
}

export type ProgramInputLevel = typeof ProgramInputLevel[keyof typeof ProgramInputLevel];


export const ProgramInputLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface ProgramInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  level: ProgramInputLevel;
  /** @minimum 1 */
  weeks: number;
  /** @minimum 1 */
  sessionsPerWeek: number;
}

export interface ClassSession {
  id: number;
  name: string;
  coach: string;
  date: string;
  startTime: string;
  durationMinutes: number;
  attendees: number;
  capacity: number;
  category: string;
  color: string;
}

export interface ClassInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  coach: string;
  /** @minLength 1 */
  date: string;
  /** @minLength 1 */
  startTime: string;
  /** @minimum 15 */
  durationMinutes: number;
  /** @minimum 1 */
  capacity: number;
  /** @minLength 1 */
  category: string;
}

export interface Checkin {
  id: number;
  memberId: number;
  memberName: string;
  checkedInAt: string;
  /** @nullable */
  className: string | null;
}

export interface CheckinInput {
  memberId: number;
  className?: string;
}

export type InsightPriority = typeof InsightPriority[keyof typeof InsightPriority];


export const InsightPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Insight {
  id: number;
  priority: InsightPriority;
  title: string;
  summary: string;
  action: string;
  /** @nullable */
  metric: string | null;
}

export type ListMembersParams = {
search?: string;
status?: ListMembersStatus;
};

export type ListMembersStatus = typeof ListMembersStatus[keyof typeof ListMembersStatus];


export const ListMembersStatus = {
  active: 'active',
  paused: 'paused',
  at_risk: 'at_risk',
  all: 'all',
} as const;