# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: FitPilot AI gym management and personal fitness API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: dashboard
  - name: members
  - name: programs
  - name: schedule
  - name: attendance
  - name: insights
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/overview:
    get:
      operationId: getDashboardOverview
      tags: [dashboard]
      summary: Get the current gym overview
      responses:
        "200":
          description: Overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverview"
  /dashboard/attendance:
    get:
      operationId: getAttendanceTrend
      tags: [dashboard]
      summary: Get recent attendance trend
      responses:
        "200":
          description: Attendance points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendancePoint"
  /activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent workspace activity
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List gym members
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, paused, at_risk, all]
            default: all
      responses:
        "200":
          description: Members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Member"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a member profile
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "404":
          description: Not found
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member profile
      parameters:
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List training programs
      responses:
        "200":
          description: Programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a training program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /classes:
    get:
      operationId: listClasses
      tags: [schedule]
      summary: List upcoming classes
      responses:
        "200":
          description: Classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSession"
    post:
      operationId: createClass
      tags: [schedule]
      summary: Create a class session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassInput"
      responses:
        "201":
          description: Created class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassSession"
  /checkins:
    get:
      operationId: listCheckins
      tags: [attendance]
      summary: List recent check-ins
      responses:
        "200":
          description: Check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkin"
    post:
      operationId: createCheckin
      tags: [attendance]
      summary: Record a check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinInput"
      responses:
        "201":
          description: Created check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
  /insights:
    get:
      operationId: listInsights
      tags: [insights]
      summary: List AI-assisted insights
      responses:
        "200":
          description: Insights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
components:
  parameters:
    MemberId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]
    DashboardOverview:
      type: object
      properties:
        memberCount: { type: integer }
        activeMembers: { type: integer }
        checkinsToday: { type: integer }
        attendanceRate: { type: number }
        attendanceChange: { type: number }
        monthlyRevenue: { type: number }
        revenueChange: { type: number }
        programCount: { type: integer }
        classCount: { type: integer }
      required: [memberCount, activeMembers, checkinsToday, attendanceRate, attendanceChange, monthlyRevenue, revenueChange, programCount, classCount]
    AttendancePoint:
      type: object
      properties:
        label: { type: string }
        checkins: { type: integer }
        capacity: { type: integer }
      required: [label, checkins, capacity]
    Activity:
      type: object
      properties:
        id: { type: integer }
        type: { type: string, enum: [checkin, signup, payment, class, program] }
        title: { type: string }
        detail: { type: string }
        timestamp: { type: string }
        memberName: { type: ["string", "null"] }
      required: [id, type, title, detail, timestamp, memberName]
    Member:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        plan: { type: string, enum: [Core, Pro, Elite] }
        status: { type: string, enum: [active, paused, at_risk] }
        joinedAt: { type: string }
        lastVisit: { type: ["string", "null"] }
        visitsThisMonth: { type: integer }
        goal: { type: string }
        avatarColor: { type: string }
      required: [id, name, email, phone, plan, status, joinedAt, lastVisit, visitsThisMonth, goal, avatarColor]
    MemberInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        phone: { type: string }
        plan: { type: string, enum: [Core, Pro, Elite] }
        goal: { type: string, minLength: 1 }
      required: [name, email, plan, goal]
    MemberUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, minLength: 3 }
        phone: { type: string }
        plan: { type: string, enum: [Core, Pro, Elite] }
        status: { type: string, enum: [active, paused, at_risk] }
        goal: { type: string, minLength: 1 }
    Program:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        level: { type: string, enum: [beginner, intermediate, advanced] }
        weeks: { type: integer }
        sessionsPerWeek: { type: integer }
        activeMembers: { type: integer }
        color: { type: string }
      required: [id, name, description, level, weeks, sessionsPerWeek, activeMembers, color]
    ProgramInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        level: { type: string, enum: [beginner, intermediate, advanced] }
        weeks: { type: integer, minimum: 1 }
        sessionsPerWeek: { type: integer, minimum: 1 }
      required: [name, description, level, weeks, sessionsPerWeek]
    ClassSession:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        coach: { type: string }
        date: { type: string }
        startTime: { type: string }
        durationMinutes: { type: integer }
        attendees: { type: integer }
        capacity: { type: integer }
        category: { type: string }
        color: { type: string }
      required: [id, name, coach, date, startTime, durationMinutes, attendees, capacity, category, color]
    ClassInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        coach: { type: string, minLength: 1 }
        date: { type: string, minLength: 1 }
        startTime: { type: string, minLength: 1 }
        durationMinutes: { type: integer, minimum: 15 }
        capacity: { type: integer, minimum: 1 }
        category: { type: string, minLength: 1 }
      required: [name, coach, date, startTime, durationMinutes, capacity, category]
    Checkin:
      type: object
      properties:
        id: { type: integer }
        memberId: { type: integer }
        memberName: { type: string }
        checkedInAt: { type: string }
        className: { type: ["string", "null"] }
      required: [id, memberId, memberName, checkedInAt, className]
    CheckinInput:
      type: object
      properties:
        memberId: { type: integer }
        className: { type: string }
      required: [memberId]
    Insight:
      type: object
      properties:
        id: { type: integer }
        priority: { type: string, enum: [high, medium, low] }
        title: { type: string }
        summary: { type: string }
        action: { type: string }
        metric: { type: ["string", "null"] }
      required: [id, priority, title, summary, action, metric]