# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FitPilot AI gym management and personal fitness API
 * OpenAPI spec version: 0.2.0
 */
import type { ActivityType } from './activityType';

export interface Activity {
  id: number;
  type: ActivityType;
  title: string;
  detail: string;
  timestamp: string;
  /** @nullable */
  memberName: string | null;
}