# lib/api-zod/src/generated/types/attendancePoint.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FitPilot AI gym management and personal fitness API
 * OpenAPI spec version: 0.2.0
 */

export interface AttendancePoint {
  label: string;
  checkins: number;
  capacity: number;
}