# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FitPilot AI gym management and personal fitness API
 * OpenAPI spec version: 0.2.0
 */

export interface Checkin {
  id: number;
  memberId: number;
  memberName: string;
  checkedInAt: string;
  /** @nullable */
  className: string | null;
}