# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FitPilot AI gym management and personal fitness API
 * OpenAPI spec version: 0.2.0
 */

export interface HealthStatus {
  status: string;
}