# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FitPilot AI gym management and personal fitness API
 * OpenAPI spec version: 0.2.0
 */
import type { MemberUpdatePlan } from './memberUpdatePlan';
import type { MemberUpdateStatus } from './memberUpdateStatus';

export interface MemberUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 3 */
  email?: string;
  phone?: string;
  plan?: MemberUpdatePlan;
  status?: MemberUpdateStatus;
  /** @minLength 1 */
  goal?: string;
}