# lib/api-zod/src/generated/types/programInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * FitPilot AI gym management and personal fitness API
 * OpenAPI spec version: 0.2.0
 */
import type { ProgramInputLevel } from './programInputLevel';

export interface ProgramInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  level: ProgramInputLevel;
  /** @minimum 1 */
  weeks: number;
  /** @minimum 1 */
  sessionsPerWeek: number;
}